requirements.txt
pandas
pyarrow
duckdb
python-dotenv
kaggle
jupyter
